service/requirements.txt
Django==1.10.5
django-cors-headers==2.0.0
django-guardian==1.4.6
djangorestframework==3.5.3
djangorestframework-jsonapi==2.1.1
drf-nested-routers==0.11.1
inflection==0.3.1
psycopg2==2.6.2
six==1.10.0
